tools/app/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.1.0"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "tabulate>=0.9",
]

[project.scripts]
app = "app.main:cli"

[tool.setuptools.packages.find]
where = ["src"]